requirements.txt

streamlit==1.46.0
langchain==0.3.25
langchain-community==0.3.25
chromadb==1.0.12
sentence-transformers==4.1.0
python-dotenv==1.1.0
tiktoken==0.9.0
PyPDF2==3.0.1
pandas==2.3.0
numpy==2.3.0
transformers==4.52.4
accelerate==1.7.0
bitsandbytes==0.41.3
llama-cpp-python==0.3.9
torch>=2.7.1